pyetf/requirements.txt
aiohttp==3.8.4
beautifulsoup4==4.12.0
mysql_connector_repackaged==0.3.1
pyetfdb_scraper==0.1.8
pytest==7.1.2
requests==2.28.1
setuptools==63.4.1
